DFT Regeneration and test for spw_spw_escChar_tch_loopback_sam_rh71_ek MH3-89785

+++ apps/spw/spw_escChar_tch_loopback/firmware/src/config/sam_rh71_ek/peripheral/spw/plib_spw_pktrx.h
@@ -66,7 +66,7 @@
 // *****************************************************************************
 // *****************************************************************************
 /* Return count field value in SPW PKTRX previous status word. This is the number of packets received in previous buffer. */
-#define SPW_PKTRX_PREV_STATUS_GET_COUNT(status) ( ((status) & SPW_PKTRX_PREV_STATUS_CNT_MASK) >> SPW_PKTRX1_PREVBUFSTS_CNT_Pos )
+#define SPW_PKTRX_PREV_STATUS_GET_COUNT(status) (uint16_t)( ((status) & SPW_PKTRX_PREV_STATUS_CNT_MASK) >> SPW_PKTRX1_PREVBUFSTS_CNT_Pos )
 
 // *****************************************************************************
 // *****************************************************************************
